+++ src/main/java/org/mule/module/hubspot/client/HubSpotClientUtils.java
@@ -9,12 +9,15 @@
 
 package org.mule.module.hubspot.client;
 
+import java.io.ByteArrayInputStream;
 import java.io.IOException;
+import java.io.InputStream;
 
 import javax.ws.rs.core.MediaType;
 
+import org.codehaus.jackson.JsonGenerationException;
 import org.codehaus.jackson.JsonParseException;
-import org.codehaus.jackson.JsonParser;
+import org.codehaus.jackson.map.JsonMappingException;
 import org.codehaus.jackson.map.ObjectMapper;
 import org.mule.module.hubspot.exception.HubSpotConnectorAccessTokenExpiredException;
 import org.mule.module.hubspot.exception.HubSpotConnectorException;
@@ -25,6 +28,31 @@ import com.sun.jersey.api.client.WebResource;
 
 public class HubSpotClientUtils {
 	
+	static final private ObjectMapper jacksonMapper = new ObjectMapper();
+	
+	
+	static public String transformObjectToJson(Object o) throws HubSpotConnectorException {
+		try {
+			return jacksonMapper.writeValueAsString(o);
+		} catch (JsonGenerationException e) {
+			throw new HubSpotConnectorException("Cannot generate the Json from object", e);
+		} catch (JsonMappingException e) {
+			throw new HubSpotConnectorException("Cannot map the Json from object", e);
+		} catch (IOException e) {
+			throw new HubSpotConnectorException(e);
+		}
+	}
+	
+	static public String webResourceGet(WebResource wr, String userId, HubSpotWebResourceMethods method) 
+			throws HubSpotConnectorAccessTokenExpiredException, HubSpotConnectorException {
+		return webResourceGet(String.class, wr, userId, method); 
+	}
+	
+	static public String webResourceGet(WebResource wr, String userId, HubSpotWebResourceMethods method, String requestBody) 
+			throws HubSpotConnectorAccessTokenExpiredException, HubSpotConnectorException {
+		return webResourceGet(String.class, wr, userId, method, requestBody); 
+	}
+	
 	/**
 	 * Method used to eliminate boilerplate handling exceptions when calling get(String.class) from a WebResource
 	 * 
@@ -34,28 +62,42 @@ public class HubSpotClientUtils {
 	 * @throws HubSpotConnectorAccessTokenExpiredException If the service responded with a 401 means that the session has expired
 	 * @throws HubSpotConnectorException If is not a 401 it will throw this exception
 	 */
-	static public String webResourceGet(WebResource wr, String userId, HubSpotWebResourceMethods method) 
+	static public <T> T webResourceGet(Class<T> type, WebResource wr, String userId, HubSpotWebResourceMethods method) 
 			throws HubSpotConnectorAccessTokenExpiredException, HubSpotConnectorException {
-		return webResourceGet(wr, userId, method, null);
+		return webResourceGet(type, wr, userId, method, null);
 	}
 	
 	
-	static public String webResourceGet(WebResource wr, String userId, HubSpotWebResourceMethods method, String requestBody) 
+	@SuppressWarnings("unchecked")
+	static public <T> T webResourceGet(Class<T> type, WebResource wr, String userId, HubSpotWebResourceMethods method, String requestBody) 
 			throws HubSpotConnectorAccessTokenExpiredException, HubSpotConnectorException {
 		try {
-			return webResourceCallByEnumType(wr, method, requestBody);
+			String res = webResourceCallByEnumType(wr, method, requestBody);
+			
+			if (type.equals(String.class)) {
+				return (T)res;
+			} else {
+				InputStream is = new ByteArrayInputStream(res.getBytes());
+				return jacksonMapper.readValue(is, type);
+			}
 		} catch (UniformInterfaceException e) {
 			int statusCode = e.getResponse().getStatus();
 			
 			// The code 204 is returned as a successful operation with no response, but as the expected parameter is a String.class it throws a UniformInterfaceException.
 			if (statusCode == 204) {
-				return "";
+				return null;
 			} else if (statusCode == 401) {
-				// TODO: Add verification if it has a refresh token. If it has it call refresh process and then make again the get() call
+				// FEATURE: Add verification if it has a refresh token. If it has it call refresh process and then make again the get() call
 				throw new HubSpotConnectorAccessTokenExpiredException("The access token for the userId " + userId + "has expired", e);
 			} else {
 				throw new HubSpotConnectorException("ERROR - statusCode: " + statusCode, e);
 			}
+		} catch (JsonParseException e) {
+			throw new HubSpotConnectorException("ERROR - Error Parsing the JSON", e);
+		} catch (JsonMappingException e) {
+			throw new HubSpotConnectorException("ERROR - Error Mapping the JSON", e);
+		} catch (IOException e) {
+			throw new HubSpotConnectorException(e);
 		}
 	}
 	
@@ -71,16 +113,5 @@ public class HubSpotClientUtils {
 		} else {
 			return null;
 		}
-	}
-	
-	static public void checkValidJson(final String json) throws HubSpotConnectorException {
-	   try {
-		      final JsonParser parser = new ObjectMapper().getJsonFactory().createJsonParser(json);
-		      while (parser.nextToken() != null) {}
-		   } catch (JsonParseException jpe) {
-			   throw new HubSpotConnectorException("The contactJson contains a JSON malformed", jpe);
-		   } catch (IOException ioe) {
-			   throw new HubSpotConnectorException("The contactJson cannot be readed", ioe);
-		   }
 	}
 }